Trying `:digit:` and `-P`
So it turns out that BSD `grep` handles `\d`, but GNU `grep` doesn't. That's why it worked here on my Mac, but didn't work on the Ubuntu image in GitHub Actions. So I'm trying both the `:digit:` form of the character class, and `-P`, which apparently says to use the Perl regex patterns, which include `\d`.